+++ src/main/webapp/app/model/tool/map/PlanCourse.js
@@ -23,7 +23,7 @@ Ext.define('Ssp.model.tool.map.PlanCourse', {
              {name:'formattedCourse', type:'string'},
              {name:'courseTitle',type:'string'},
              {name:'courseDescription',type:'string'},
-             {name:'creditHours', type:'int'},
+             {name:'creditHours', type:'float'},
              {name:'orderInTerm',type:'int'},
              {name:'isDev',type:'boolean'},
              {name:'contactNotes',type:'string'},
